+++ x/vesting/keeper/integration_test.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 	"time"
 
+	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 
 	"cosmossdk.io/math"
@@ -726,8 +727,9 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 
 		// Perform clawback before cliff
 		msg := types.NewMsgClawback(funder, vestingAddr, dest)
-		_, err := s.app.VestingKeeper.Clawback(ctx, msg)
+		res, err := s.app.VestingKeeper.Clawback(ctx, msg)
 		Expect(err).To(BeNil())
+		Expect(res.Coins).To(Equal(vestingAmtTotal), "expected different coins to be clawed back")
 
 		// All initial vesting amount goes to dest
 		bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
@@ -751,6 +753,7 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 		vesting = clawbackAccount.GetVestingCoins(s.ctx.BlockTime())
 		expVestedAmount := amt.Mul(sdk.NewInt(cliff))
 		expVested := sdk.NewCoins(sdk.NewCoin(stakeDenom, expVestedAmount))
+		unvested := vestingAmtTotal.Sub(vested...)
 
 		s.Require().Equal(expVested, vested)
 		s.Require().True(expVestedAmount.GT(sdk.NewInt(0)))
@@ -764,8 +767,9 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 		// Perform clawback
 		msg := types.NewMsgClawback(funder, vestingAddr, dest)
 		ctx := sdk.WrapSDKContext(s.ctx)
-		_, err = s.app.VestingKeeper.Clawback(ctx, msg)
+		res, err := s.app.VestingKeeper.Clawback(ctx, msg)
 		Expect(err).To(BeNil())
+		Expect(res.Coins).To(Equal(unvested), "expected unvested coins to be clawed back")
 
 		bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
 		bG := s.app.BankKeeper.GetBalance(s.ctx, vestingAddr, stakeDenom)
@@ -807,8 +811,9 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 		// Perform clawback
 		msg := types.NewMsgClawback(funder, vestingAddr, dest)
 		ctx := sdk.WrapSDKContext(s.ctx)
-		_, err = s.app.VestingKeeper.Clawback(ctx, msg)
+		res, err := s.app.VestingKeeper.Clawback(ctx, msg)
 		Expect(err).To(BeNil())
+		Expect(res.Coins).To(Equal(unvested), "expected only coins to be clawed back")
 
 		bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
 		bG := s.app.BankKeeper.GetBalance(s.ctx, vestingAddr, stakeDenom)
@@ -848,8 +853,10 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 		// Perform clawback
 		msg := types.NewMsgClawback(funder, vestingAddr, dest)
 		ctx := sdk.WrapSDKContext(s.ctx)
-		_, err = s.app.VestingKeeper.Clawback(ctx, msg)
-		Expect(err).To(BeNil())
+		res, err := s.app.VestingKeeper.Clawback(ctx, msg)
+		Expect(err).To(BeNil(), "expected no error during clawback")
+		Expect(res).ToNot(BeNil(), "expected response not to be nil")
+		Expect(res.Coins).To(BeEmpty(), "expected nothing to be clawed back")
 
 		bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
 		bG := s.app.BankKeeper.GetBalance(s.ctx, vestingAddr, stakeDenom)
@@ -876,8 +883,9 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 
 		// Perform clawback before cliff - funds should go to new funder (no dest address defined)
 		msg := types.NewMsgClawback(newFunder, vestingAddr, sdk.AccAddress([]byte{}))
-		_, err = s.app.VestingKeeper.Clawback(ctx, msg)
-		s.Require().NoError(err)
+		res, err := s.app.VestingKeeper.Clawback(ctx, msg)
+		Expect(err).To(BeNil())
+		Expect(res.Coins).To(Equal(vestingAmtTotal), "expected different coins to be clawed back")
 
 		// All initial vesting amount goes to funder
 		bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
@@ -932,8 +940,9 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 
 			// Perform clawback before cliff
 			msg := types.NewMsgClawback(authtypes.NewModuleAddress(govtypes.ModuleName), vestingAddr, dest)
-			_, err := s.app.VestingKeeper.Clawback(ctx, msg)
+			res, err := s.app.VestingKeeper.Clawback(ctx, msg)
 			Expect(err).To(BeNil())
+			Expect(res.Coins).To(Equal(vestingAmtTotal), "expected different coins to be clawed back")
 
 			// All initial vesting amount goes to community pool instead of dest
 			bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
@@ -963,6 +972,7 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 			vesting = clawbackAccount.GetVestingCoins(s.ctx.BlockTime())
 			expVestedAmount := amt.Mul(sdk.NewInt(cliff))
 			expVested := sdk.NewCoins(sdk.NewCoin(stakeDenom, expVestedAmount))
+			unvested := vestingAmtTotal.Sub(vested...)
 
 			s.Require().Equal(expVested, vested)
 			s.Require().True(expVestedAmount.GT(sdk.NewInt(0)))
@@ -987,8 +997,9 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 			// Perform clawback
 			msg := types.NewMsgClawback(authtypes.NewModuleAddress(govtypes.ModuleName), vestingAddr, dest)
 			ctx := sdk.WrapSDKContext(s.ctx)
-			_, err = s.app.VestingKeeper.Clawback(ctx, msg)
+			res, err := s.app.VestingKeeper.Clawback(ctx, msg)
 			Expect(err).To(BeNil())
+			Expect(res.Coins).To(Equal(unvested), "expected unvested coins to be clawed back")
 
 			bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
 			bG := s.app.BankKeeper.GetBalance(s.ctx, vestingAddr, stakeDenom)
@@ -1044,8 +1055,9 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 			// Perform clawback
 			msg := types.NewMsgClawback(funder, vestingAddr, dest)
 			ctx := sdk.WrapSDKContext(s.ctx)
-			_, err = s.app.VestingKeeper.Clawback(ctx, msg)
+			res, err := s.app.VestingKeeper.Clawback(ctx, msg)
 			Expect(err).To(BeNil())
+			Expect(res.Coins).To(Equal(unvested), "expected only unvested coins to be clawed back")
 
 			bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
 			bG := s.app.BankKeeper.GetBalance(s.ctx, vestingAddr, stakeDenom)
@@ -1096,8 +1108,9 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 			// Perform clawback
 			msg := types.NewMsgClawback(authtypes.NewModuleAddress(govtypes.ModuleName), vestingAddr, dest)
 			ctx := sdk.WrapSDKContext(s.ctx)
-			_, err = s.app.VestingKeeper.Clawback(ctx, msg)
-			Expect(err).To(BeNil())
+			res, err := s.app.VestingKeeper.Clawback(ctx, msg)
+			Expect(err).To(BeNil(), "expected no error during clawback")
+			Expect(res.Coins).To(BeEmpty(), "expected nothing to be clawed back after end of vesting schedules")
 
 			bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
 			bG := s.app.BankKeeper.GetBalance(s.ctx, vestingAddr, stakeDenom)
@@ -1129,8 +1142,9 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 
 			// Perform clawback before cliff - funds should go to new funder (no dest address defined)
 			msg := types.NewMsgClawback(authtypes.NewModuleAddress(govtypes.ModuleName), vestingAddr, nil)
-			_, err = s.app.VestingKeeper.Clawback(ctx, msg)
-			s.Require().NoError(err)
+			res, err := s.app.VestingKeeper.Clawback(ctx, msg)
+			Expect(err).To(BeNil())
+			Expect(res.Coins).To(Equal(vestingAmtTotal), "expected different coins to be clawed back")
 
 			// All initial vesting amount goes to funder
 			bF := s.app.BankKeeper.GetBalance(s.ctx, funder, stakeDenom)
@@ -1156,7 +1170,7 @@ var _ = Describe("Clawback Vesting Accounts - claw back tokens", func() {
 			msg := types.NewMsgClawback(authtypes.NewModuleAddress(govtypes.ModuleName), vestingAddr, dest)
 			_, err := s.app.VestingKeeper.Clawback(s.ctx, msg)
 			Expect(err).To(HaveOccurred(), "expected error")
-			Expect(err.Error()).To(ContainSubstring("doesn't have governance clawback enabled"))
+			Expect(err.Error()).To(ContainSubstring("%s: account does not have governance clawback enabled", vestingAddr.String()))
 		})
 	})
 })
@@ -1232,7 +1246,7 @@ var _ = Describe("Clawback Vesting Account - Barberry bug", func() {
 		Expect(ok).To(BeTrue(), "account is not a clawback vesting account")
 	})
 
-	Context("when creating a clawback vesting account", func() {
+	Context("when funding a clawback vesting account", func() {
 		testcases := []struct {
 			name         string
 			lockupCoins  sdk.Coins
@@ -1260,7 +1274,7 @@ var _ = Describe("Clawback Vesting Account - Barberry bug", func() {
 				name:        "fail - negative amounts for the lockup period",
 				lockupCoins: coinsWithNegAmount,
 				expError:    true,
-				errContains: "invalid amount in lockup periods, amounts must be positive",
+				errContains: errortypes.ErrInvalidCoins.Wrap(coinsWithNegAmount.String()).Error(),
 			},
 			{
 				name:         "fail - negative amounts for the vesting period",
@@ -1272,7 +1286,7 @@ var _ = Describe("Clawback Vesting Account - Barberry bug", func() {
 				name:        "fail - zero amount for the lockup period",
 				lockupCoins: coinsWithZeroAmount,
 				expError:    true,
-				errContains: "invalid amount in lockup periods, amounts must be positive",
+				errContains: errortypes.ErrInvalidCoins.Wrap(coinsWithZeroAmount.String()).Error(),
 			},
 			{
 				name:         "fail - zero amount for the vesting period",